# api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Qlxion API
  description: |
    Qlxion API Documentation - API Gateway dan Auth Service.
    
    Semua endpoint yang dilindungi (Bearer Auth) memerlukan header:
    `Authorization: Bearer <access_token>`
    
    Gateway akan meneruskan claims JWT ke backend services melalui header:
    - `X-User-Id`: UUID user
    - `X-Tenant-Id`: UUID tenant
    - `X-User-Email`: Email user
    - `X-User-Roles`: Comma-separated roles
  version: 1.0.0
  contact:
    name: Qlxion Team
    email: dev@qlxion.com
servers:
  - url: http://localhost:8000
    description: Local development (API Gateway)
  - url: http://localhost:8001
    description: Local development (Auth Service direct)
security:
  - BearerAuth: []
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: List users
      description: Get paginated list of users
      parameters:
        - name: tenant_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: integer
            enum: [0, 1, 2]
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserResponse'

    post:
      tags:
        - Users
      summary: Create user
      description: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user
      description: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - Users
      summary: Update user
      description: Update user information
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated

    delete:
      tags:
        - Users
      summary: Delete user
      description: Soft delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted
  /users/{id}/roles:
    get:
      tags:
        - Users
      summary: Get user roles
      description: Get roles assigned to a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'

    post:
      tags:
        - Users
      summary: Assign role
      description: Assign a role to a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Role assigned

  # ============================================
  # Roles
  # ============================================
...
  /tenants:
    get:
      tags:
        - Tenants
      summary: List tenants
      parameters:
        - name: status
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of tenants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTenantResponse'

    post:
      tags:
        - Tenants
      summary: Create tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Tenant created
  /tenants/{id}:
    get:
      tags:
        - Tenants
      summary: Get tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tenant found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'

    put:
      tags:
        - Tenants
      summary: Update tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTenantRequest'
      responses:
        '200':
          description: Tenant updated

    delete:
      tags:
        - Tenants
      summary: Delete tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Tenant deleted

  # ============================================
  # Sessions
  # ============================================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ============================================
  # Schemas
  # ============================================
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        full_name:
          type: string
        email:
          type: string
          format: email
        username:
          type: string
        phone:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        is_email_verified:
          type: boolean
        is_phone_verified:
          type: boolean
        status:
          type: integer
          enum: [0, 1, 2]
          description: 0=BLOCKED, 1=ACTIVE, 2=PENDING
        last_login_at:
          type: string
          format: date-time
          nullable: true
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        profile:
          $ref: '#/components/schemas/Profile'
        created_at:
          type: string
          format: date-time
    Profile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        avatar:
          type: string
          nullable: true
        full_name:
          type: string
          nullable: true
        nik_number:
          type: string
          nullable: true
        place_birth:
          type: string
          nullable: true
        date_of_birth:
          type: string
          format: date
          nullable: true
        mobile:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        zip_code:
          type: string
          nullable: true
        extras:
          type: object
    Role:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
          nullable: true
        code:
          type: string
          example: ROLE_ADMIN
        name:
          type: string
          example: Administrator
        description:
          type: string
          nullable: true
        is_system_defined:
          type: boolean
        status:
          type: integer
          enum: [0, 1]
          description: 0=INACTIVE, 1=ACTIVE
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
        created_at:
          type: string
          format: date-time
    Permission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        resource:
          type: string
          example: user
        action:
          type: string
          example: create
        description:
          type: string
          nullable: true
        is_system_defined:
          type: boolean
        created_at:
          type: string
          format: date-time
    Tenant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          example: PTA
        name:
          type: string
          example: Perusahaan A
        domain:
          type: string
          nullable: true
        config:
          type: object
        status:
          type: integer
          enum: [0, 1, 2]
          description: 0=INACTIVE, 1=ACTIVE, 2=SUSPENDED
        created_at:
          type: string
          format: date-time
...
    CreateUserRequest:
      type: object
      required:
        - tenant_id
        - full_name
        - email
        - username
      properties:
        tenant_id:
          type: string
          format: uuid
        full_name:
          type: string
        email:
          type: string
          format: email
        username:
          type: string
        password:
          type: string
        phone:
          type: string
          nullable: true
        status:
          type: integer
          default: 1
        role_ids:
          type: array
          items:
            type: string
            format: uuid
    UpdateUserRequest:
      type: object
      properties:
        full_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        status:
          type: integer
        profile:
          $ref: '#/components/schemas/Profile'
...
    CreateTenantRequest:
      type: object
      required:
        - code
        - name
      properties:
        code:
          type: string
          example: PTA
        name:
          type: string
        domain:
          type: string
          nullable: true
        config:
          type: object
    UpdateTenantRequest:
      type: object
      properties:
        name:
          type: string
        domain:
          type: string
          nullable: true
        config:
          type: object
        status:
          type: integer

    # Response wrappers
    PaginatedUserResponse:
      allOf:
        - $ref: '#/components/schemas/JSendSuccess'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/User'
            meta:
              $ref: '#/components/schemas/PaginationMeta'
...
    PaginatedTenantResponse:
      allOf:
        - $ref: '#/components/schemas/JSendSuccess'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Tenant'
            meta:
              $ref: '#/components/schemas/PaginationMeta'
...
    JSendSuccess:
      type: object
      properties:
        status:
          type: string
          example: success
        timestamp:
          type: string
          format: date-time
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer

  # ============================================
  # Responses
  # ============================================
  responses:
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: error
              code:
                type: string
                example: NOT_FOUND
              message:
                type: string
...